Moved makeExplanationOrSuggestion() from AxiosStatus to AxiosMonitor.
